requirements.txt
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.4
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
Faker==40.1.2
Flask==2.2.5
Flask-Bcrypt==1.0.1

fonttools==4.61.1
freetype-py==2.5.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
multidict==6.7.0
mysql-connector-python==8.2.0

nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.21.12
psutil==7.2.1
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyMySQL==1.1.0
pyparsing==3.3.1
pypdf==6.6.2
PySocks==1.7.1
pytest==9.0.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==1.0.2
python-telegram-bot==21.7
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
reportlab==4.4.9
requests==2.31.0
rlPyCairo==0.4.0
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
svglib==1.6.0
tinycss2==1.5.1
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.0
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.0.1
wsproto==1.3.2
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
yarl==1.22.0
Flask-WTF
Flask-MySQLdb
mysqlclient
pymysql
